Extract account details card empty state
