+++ src/app/pages/ConsensusAccountDetailsPage/Staking.tsx
@@ -1,14 +1,10 @@
 import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
-import Box from '@mui/material/Box'
 import Card from '@mui/material/Card'
 import CardHeader from '@mui/material/CardHeader'
 import CardContent from '@mui/material/CardContent'
 import Link from '@mui/material/Link'
 import Skeleton from '@mui/material/Skeleton'
-import Typography from '@mui/material/Typography'
-import { styled } from '@mui/material/styles'
-import StackedBarChartIcon from '@mui/icons-material/StackedBarChart'
 import { COLORS } from '../../../styles/theme/colors'
 import { Account, useGetConsensusAccountsAddressDelegations } from '../../../oasis-nexus/api'
 import { useRequiredScopeParam } from '../../../app/hooks/useScopeParam'
@@ -17,6 +13,8 @@ import { NUMBER_OF_ITEMS_ON_DASHBOARD as PAGE_SIZE } from '../../config'
 import { useSearchParamsPagination } from '../../components/Table/useSearchParamsPagination'
 import { Delegations } from '../..//components/Delegations'
 import { docs, wallet } from '../../utils/externalLinks'
+import { t } from 'i18next'
+import { ConsensusAccountCardEmptyState } from './ConsensusAccountCardEmptyState'
 
 type StakingProps = {
   account: Account | undefined
@@ -76,7 +74,13 @@ const StakingContent: FC<StakingContentProps> = ({ address }) => {
   }
 
   if (isFetched && !delegationsQuery.data?.data.delegations.length) {
-    return <StakingEmptyState />
+    return (
+      <ConsensusAccountCardEmptyState label={t('account.notStaking')}>
+        <Link href={wallet.homepage} rel="noopener noreferrer" target="_blank">
+          {t('account.startStaking')}
+        </Link>
+      </ConsensusAccountCardEmptyState>
+    )
   }
 
   return (
@@ -96,42 +100,5 @@ const StakingContent: FC<StakingContentProps> = ({ address }) => {
         />
       )}
     </>
-  )
-}
-
-const StyledBox = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-  [theme.breakpoints.down('sm')]: {
-    paddingTop: theme.spacing(3),
-  },
-  [theme.breakpoints.up('sm')]: {
-    paddingTop: theme.spacing(6),
-  },
-}))
-
-const StakingEmptyState: FC = () => {
-  const { t } = useTranslation()
-
-  return (
-    <StyledBox gap={3}>
-      <StackedBarChartIcon sx={{ color: COLORS.grayMedium, fontSize: 40, opacity: 0.5 }} />
-      <Typography
-        sx={{
-          color: COLORS.grayMedium,
-          fontWeight: 700,
-          maxWidth: '160px',
-          textAlign: 'center',
-          opacity: 0.5,
-        }}
-      >
-        {t('account.notStaking')}
-      </Typography>
-      <Link href={wallet.homepage} rel="noopener noreferrer" target="_blank">
-        {t('account.startStaking')}
-      </Link>
-    </StyledBox>
   )
 }
